iQIcBAABCAAGBQJYXBAfAAoJEJ5mu3zHWF1PXZEP/12NeFWqjtVCGdCtg9VwJTTD  mEe/p6bcN3WTx1rP/9024AM7G42oJAjgsOfVt4FkPVqqyPLeV1mxtGonFtmyJYag  +MtIhIa90pO3Mv8xDC3uKgAbtHhAtSZ/i0Aqpkn8Wz+bLyDzq2Ps9gVFN5wQCCi9  dh2AzQzb04UvzLgmE7xpnvxQa8LXpm5pwMh2wNorCC+vpH3SVrRNSpQwoQ222VjB  86OQVLa6iA4hom1y98I7s3A7f27EJDf1qyhyRb7rZZGuQLnNcC9Uze7EzOaIcO22  P0yIHv1+amC2WCoYkUjvqexedWUmMownbBGhBwSal7PDsRaYpH3J17Ms6O7uZ+bT  Ag0exscnhGb/ml8A028Em4y99XvJXDl5Afda4CQaGbVdXkjES9IJO2zVpP3iOEys  QZ2PTPW+3pgWyq1LXNv+iGfiEK0CUC2R4C7hx001xthYMtONAF5Ndp/tJGFua2HZ  /saoujRzOU2dy1B3br5AJlKgt5z3OfqCK4Y/aUWNVILZluDa+UwQHbixGfcd2qwT  54Iwl8YZQgfl6jE5fX9TH1XmBFCiwD/dOzYYbvewvsuwk/3Zga6I2O84xoeNXB8l  d5kIrbxjRx0vB1yZ0GecvyuKyoOjOr6Tq0tXk9rqhaOTL9CPGyzmlBDyOk2Hry30  Qda8BiNi8+UjRHBTlrro  =1OXC  -----END PGP SIGNATURE-----
Keep rolling lists of rendered instances and shallow renderers.

UIP-1296
When we render components we leave them to sit around and never unmount them. This was causing some major perf regression in tests.

TO combat this we should only keep the latest rendered instances and shallow renderers mounted. Defaulted to 5 but allowing for that number to be customizable.