Don't generate exported accessor classes
Generated accessor mixins that are common to all components (i.e. they
come along via `extends UiProps`) should not be generated for every
consumer component. This code adds an annotation to signify that the
props mixin class should be exported and not generated in the consumer
code. Also modifies the builder to not generate the accessors mixins
when that annotation is present. TODO is to write a new builder or
modify the existing builder to generate these accessors mixins