Remove usage of source_gen to avoid parsing elements:
The PartBuilder util from the source_gen package always retrieves the
LibraryElement for each file that is a library (i.e. not a part file).
This requires that the analysis session resolves the entire element tree
for that file, which is expensive but more importantly results in reads
on all transitively imported files. These reads are then stored in the
asset graph as inputs for that build step, which consequently means that
the over_react builder will be re-run on many files unnecessarily because
the inputs have changed.

The over_react builder only needs the AST for each file, so the solution
here is to replace our usage of PartBuilder with a custom Builder that
manually reads the library source for each BuildStep and calls the
analyzer API `parseCompilationUnit()` to get the info needed to generate
the `.over_react.g.dart` part files.

In preliminary testing, we've seen this approach reduce the rebuild time
from ~1 minute to ~7 seconds when changing a single file in a large
library that has many transitive imports.