Promote missing over_react part to a SEVERE log.
This will cause the build to fail, which is a breaking change as
previously the builder would only warn. This change is being made
because the part directive is required to function correctly on
Dart 2 (without it, the component will fail at runtime). In other
words, we are promoting the failure from runtime to buildtime.