Only test a single error/fix at a time for less ambiguity and better mismatch messages, clean up tests
Tests normally want to verify a specific error or fix, and not multiple,
so having the matchers match any of the errors or fixes didn't make sense.

anyOf doesn't provide very good mismatch messages, so getting rid of
those in the error/fix matchers improves the readability of test
failures by quite a bit.