Remove unnecessary assert as a workaround for false positive with Fragment
isPotentiallyValidComponentType is incorrect in dart2js (see TODO on that function)
and it returns false for Fragment's Symbol type.

There's not a great way to fix that behavior until Dart 3 (see TODO), so in the meantime,
we'll just remove this assert.

The assert really isn't necessary, since only types from ReactElements and
ReactComponentFactoryProxys get passed into it, as opposed to consumer arguments,
so we should be able to safely remove it. Any cases it might have caught should also be
covered by  test coverage on type checking code.