Fix legacy boilerplate generated mixin errors on language versions >=3.0
Example of error:

    example/boilerplate_versions/dart2_only/basic_component2.over_react.g.dart:120:44: Error: The class '_$BasicPropsAccessorsMixin' can't be used as a mixin because it isn't a mixin class nor a mixin.
      class BasicProps extends _$BasicProps with _$BasicPropsAccessorsMixin {
                                                 ^

For some reason, this is coming up in master CI in Dart 3, even though
we have a pubspec.yaml language version of 2.19.
This Seems like a build package bug, which I plan on reporting.

Since we'll need to fix this issue anyway, fix it now as a workaround.